Add files via upload
The client file was working fine as it is. We were getting memory shortage issues when we tried to run the server code on the sky mote, so we removed most of the resources except for hello resource and compiled the code. It got compiled successfully and we simulated it on Cooja as well